supabase/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
SQLAlchemy==2.0.36
asyncpg==0.30.0
supabase==2.6.0
pydantic==2.9.2